Fixed overflow when creating hash
This fixes an overflow when creating the hash for type Int64 (long). The
result is calculated before the cast, to ensure that the resulting
integer isn't negative before the modulo is applied, which can occur
when the value is sufficiently large..

By downcasting the Int64 before the modulo operation, significant
information is lost, the value will likely not result in the expected
index, resulting in an erroneous letter being pulled from the alphabet,
or an IndexOutOfRangeException being thrown due to negative values.