alphabet: handle large ratio of alphabet to seps
When using a custom alphabet, there may be a large ratio of alphabet
characters to separator characters.  When the ratio is larger than 3.5,
the alphabet and separators are adjusted.

However, if we use integer arithmetic, we may not do the adjustment when
the ratio is > 3.5 but < 4.  Use floating point arithmetic to do this
comparison.

Include a test where the ratio of alphabet characters to separator
characters is 3.714.